+++ ledger/ledger-api-common/src/main/scala/com/digitalasset/ledger/api/tls/TlsConfiguration.scala
@@ -5,35 +5,14 @@ package com.daml.ledger.api.tls
 
 import com.daml.ledger.api.tls.TlsVersion.{TlsVersion, V1, V1_1, V1_2, V1_3}
 import io.grpc.netty.GrpcSslContexts
-import io.netty.buffer.ByteBufAllocator
 import io.netty.handler.ssl.{ClientAuth, SslContext}
 import org.slf4j.LoggerFactory
 
 import java.io.{ByteArrayInputStream, File, FileInputStream, InputStream}
 import java.lang
 import java.nio.file.Files
-import javax.net.ssl.SSLEngine
 import scala.jdk.CollectionConverters._
 import scala.util.control.NonFatal
-
-final case class TlsInfo(
-    enabledCipherSuites: Seq[String],
-    enabledProtocols: Seq[String],
-    supportedCipherSuites: Seq[String],
-    supportedProtocols: Seq[String],
-)
-
-object TlsInfo {
-  def fromSslContext(sslContext: SslContext): TlsInfo = {
-    val engine: SSLEngine = sslContext.newEngine(ByteBufAllocator.DEFAULT)
-    TlsInfo(
-      enabledCipherSuites = engine.getEnabledCipherSuites.toIndexedSeq,
-      enabledProtocols = engine.getEnabledProtocols.toIndexedSeq,
-      supportedCipherSuites = engine.getSupportedCipherSuites.toIndexedSeq,
-      supportedProtocols = engine.getSupportedProtocols.toIndexedSeq,
-    )
-  }
-}
 
 final case class TlsConfiguration(
     enabled: Boolean,
@@ -72,19 +51,25 @@ final case class TlsConfiguration(
     } else None
   }
 
-  /** If enabled and all required fields are present, it returns an SslContext suitable for server usage */
+  /** If enabled and all required fields are present, it returns an SslContext suitable for server usage
+    *
+    *  Details:
+    *  We create two instances of sslContext:
+    *  1) The first one with default protocols: in order to query it for a set of supported protocols.
+    *  2) The second one with a custom set of protocols to enable.
+    *     We have used previously obtained set of supported protocols to make sure every protocol we want to enable is supported.
+    *     @see [[javax.net.ssl.SSLEngine#setEnabledProtocols]]
+    */
   def server: Option[SslContext] =
     if (enabled) {
-
       val tlsInfo = scala.util.Using.resources(
         keyCertChainInputStreamOrFail,
         keyInputStreamOrFail,
       ) { (keyCertChain: InputStream, key: InputStream) =>
-        val protocols = null.asInstanceOf[lang.Iterable[String]]
         val defaultSslContext = buildServersSslContext(
           keyCertChain = keyCertChain,
           key = key,
-          protocols = protocols,
+          protocols = null.asInstanceOf[lang.Iterable[String]],
         )
         TlsInfo.fromSslContext(defaultSslContext)
       }
@@ -93,11 +78,10 @@ final case class TlsConfiguration(
         keyCertChainInputStreamOrFail,
         keyInputStreamOrFail,
       ) { (keyCertChain: InputStream, key: InputStream) =>
-        val protocols = protocolsNames(tlsInfo)
         val sslContext = buildServersSslContext(
           keyCertChain = keyCertChain,
           key = key,
-          protocols = protocols,
+          protocols = filterSupportedProtocols(tlsInfo),
         )
         logTlsProtocolsAndCipherSuites(sslContext, isServer = true)
         Some(sslContext)
@@ -129,22 +113,22 @@ final case class TlsConfiguration(
       sslContext: SslContext,
       isServer: Boolean,
   ): Unit = {
-    val who = if (isServer) "server" else "client"
+    val who = if (isServer) "Server" else "Client"
     val tlsInfo = TlsInfo.fromSslContext(sslContext)
-    logger.info(s"TLS $who: Enabled.")
-    logger.debug(s"TLS $who: Supported protocols: ${tlsInfo.supportedProtocols.mkString(", ")}.")
-    logger.info(s"TLS $who: Enabled protocols: ${tlsInfo.enabledProtocols.mkString(", ")}.")
+    logger.info(s"$who TLS - enabled.")
+    logger.debug(s"$who TLS - supported protocols: ${tlsInfo.supportedProtocols.mkString(", ")}.")
+    logger.info(s"$who TLS - enabled protocols: ${tlsInfo.enabledProtocols.mkString(", ")}.")
     logger.debug(
-      s"TLS $who: Supported cipher suites: ${tlsInfo.supportedCipherSuites.mkString(", ")}."
+      s"$who TLS $who - supported cipher suites: ${tlsInfo.supportedCipherSuites.mkString(", ")}."
     )
-    logger.info(s"TLS $who: Enabled cipher suites: ${tlsInfo.enabledCipherSuites.mkString(", ")}.")
+    logger.info(s"$who TLS - enabled cipher suites: ${tlsInfo.enabledCipherSuites.mkString(", ")}.")
   }
 
   /** This is a side-effecting method. It modifies JVM TLS properties according to the TLS configuration. */
   def setJvmTlsProperties(): Unit =
     if (enabled && enableCertRevocationChecking) OcspProperties.enableOcsp()
 
-  private[tls] def protocolsNames(tlsInfo: TlsInfo): java.lang.Iterable[String] = {
+  private[tls] def filterSupportedProtocols(tlsInfo: TlsInfo): java.lang.Iterable[String] = {
     minimumServerProtocolVersion match {
       case None => null
       case Some(tlsVersion) =>
